% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo.bgp.r
\name{algo.bgp}
\alias{algo.bgp}
\title{Baragatti Grimaud and Pommeret MCMC chain}
\usage{
algo.bgp(evals, chains, last, cfg, pdesc, priv)
}
\description{
Baragatti Grimaud and Pommeret MCMC chain
}
\seealso{
Other algos: \code{\link{algo.diffevo}},
  \code{\link{algo.mh}}, \code{\link{algo.wl}}
}

