% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo.diffevolution.r
\name{algo.diffevo}
\alias{algo.diffevo}
\title{differential evolution}
\usage{
algo.diffevo(evals, chains, last, cfg, pdesc, priv)
}
\description{
differential evolution
}
\seealso{
Other algos: \code{\link{algo.bgp}}, \code{\link{algo.mh}},
  \code{\link{algo.wl}}
}

