% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/algo.gibbs.r
\name{algo.gibbs}
\alias{algo.gibbs}
\title{Metropolis Hasting MCMC chain updating}
\usage{
algo.gibbs(rd, param_data, niter, cf, pdesc, priv)
}
\description{
Metropolis Hasting MCMC chain updating
}
\seealso{
Other algos: \code{\link{algo.bgp}};
  \code{\link{algo.diffevo}}; \code{\link{algo.mh}};
  \code{\link{algo.wl}}
}

