% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo.mh.r
\name{algo.mh}
\alias{algo.mh}
\title{Metropolis Hasting MCMC chain updating}
\usage{
algo.mh(param_data, rd, cf, N, niter)
}
\description{
Metropolis Hasting MCMC chain updating
}
\seealso{
Other algos: \code{\link{algo.bgp}},
  \code{\link{algo.diffevo}}, \code{\link{algo.gibbs}},
  \code{\link{algo.slicemax}}, \code{\link{algo.wl}}
}

