% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo.slicemax.R
\name{algo.slicemax}
\alias{algo.slicemax}
\title{Simple aglorithm which compute a slice, then takes the minimum}
\usage{
algo.slicemax(rd, param_data, niter, cf, pdesc, priv)
}
\description{
Simple aglorithm which compute a slice, then takes the minimum
}
\seealso{
Other algos: \code{\link{algo.bgp}},
  \code{\link{algo.diffevo}}, \code{\link{algo.gibbs}},
  \code{\link{algo.mh}}, \code{\link{algo.wl}}
}

