\name{algo.slices}
\alias{algo.slices}
\title{Compute the objective function on a grid}
\usage{
  algo.slices(mcf, ns = 30, pad = 0.2)
}
\description{
  Will vary each parameter independently from the starting
  value and store all values. By default it will run for
  all parameters, otherwise it uses the argument list
}
\seealso{
  Other algos: \code{\link{algo.bgp}},
  \code{\link{algo.mh}}, \code{\link{algo.wl}}
}

