% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo.wl.r
\name{algo.wl}
\alias{algo.wl}
\title{Wang-Landau MCMC}
\usage{
algo.wl(param_data, rd, cf, N, niter)
}
\description{
Multi chain MCMC based on Wang-Landau
}
\seealso{
Other algos: \code{\link{algo.bgp}},
  \code{\link{algo.diffevo}}, \code{\link{algo.gibbs}},
  \code{\link{algo.mh}}, \code{\link{algo.slicemax}}
}

