% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.econmetrics.r
\name{estimate.mopt_env}
\alias{estimate.mopt_env}
\title{takes the best value, compute the score and
the standard error using the sandwich formula}
\usage{
estimate.mopt_env(me)
}
\description{
takes the best value, compute the score and
the standard error using the sandwich formula
}

