\name{melt.mopt_env}
\alias{melt.mopt_env}
\title{melts data with only sampled parameters
if set cols='m' or cols='p' extract the wide format with
only parameters or only moments}
\usage{
  melt.mopt_env(me, cols = NULL)
}
\description{
  melts data with only sampled parameters if set cols='m'
  or cols='p' extract the wide format with only parameters
  or only moments
}

