% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mopt-package.r
\docType{package}
\name{mopt}
\alias{mopt}
\alias{mopt-package}
\title{Parallel optimizer for R}
\description{
implements MCMC in parallel on OpenMP or MPI
}
\details{
The main goal of this package is to provide a library for simulated moment
estimation done in parallel on either MPI or openMP. The code is writtern 
specificaly for moment estiamtion in the sense that the objective function
should return, not only an overall value, but also each moments. This is 
then used in the analysis part. For instance the researcher can analyze
how each moment is affected by each parameter.

The main function is \code{\link{runMOpt}}, and the associated \code{\link{mopt_config}}. This gives
full control to the researchers of the methods used to look for the minimum distance estimator. Several
algorithms are implemented and we encourage other people to write more of them and leverage the parallelization
structure built into this library.

A very useful function is \code{\link{compute.slices}} and the associated \code{\link{plot.slices}} which will generate plots of how
the different moments vary as you vary each parameters separately.
}
\author{
Ran Gu \email{lionup@gmail.com}, Thibaut Lamadon \email{thibaut.lamadon@gmail.com} , Florian Oswald \email{florian.oswald@gmail.com}
}
\references{
\url{https://github.com/tlamadon/mopt}
}
\seealso{
runMOpt
}

