% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inc.mopt.r
\name{mopt_obj_wrapper}
\alias{mopt_obj_wrapper}
\title{wraper for the objective function. Made public for multicore}
\usage{
mopt_obj_wrapper(p, objfunc = NA, errfile = "param_error.dat")
}
\description{
wraper for the objective function. Made public for multicore
}

