% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc.reporting.r
\name{plot.mopt_env}
\alias{plot.mopt_env}
\title{generates different plots from a chain}
\usage{
\method{plot}{mopt_env}(me, what = "na", select = "untempered",
  varblack = c(), taildata = 0)
}
\description{
generates different plots from a chain
}

