\name{predict.mopt_config}
\alias{predict.mopt_config}
\title{return some versions of the parameters}
\usage{
  predict.mopt_config(cf, what = "p.all", base = "")
}
\arguments{
  \item{what}{can be 'p.all' the best parameter set as a
  list, 'p.sd' for sampled parameters with standard
  deviations based on coldest chain, 'm' for list of
  simulated and data moments next to each other}
}
\description{
  return some versions of the parameters
}

