% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inc.reporting.r
\name{predict.mopt_env}
\alias{predict.mopt_env}
\title{return some versions of the parameters}
\usage{
\method{predict}{mopt_env}(me, what = "p.all", base = "", sort = FALSE)
}
\arguments{
\item{what}{can be 'p.all' the best parameter set as a list, 'p.sd' for
sampled parameters with standard deviations based on coldest chain, 'm' for list of
simulated and data moments next to each other}
}
\description{
return some versions of the parameters
}

