% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc.mopt.r
\name{prepare.mopt_config}
\alias{prepare.mopt_config}
\title{prepares mopt to run with either MPI or openMP or just serial}
\usage{
prepare.mopt_config(cf)
}
\description{
prepares mopt to run with either MPI or openMP or just serial
}

