% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc.mopt.r
\name{rollcall}
\alias{rollcall}
\title{cluster rollcall}
\usage{
rollcall(dir)
}
\arguments{
\item{dir}{path to the log directory}
}
\description{
cluster reporting function. makes every worker state their name and 
writes into a log file
}

