\name{score.mopt_env}
\alias{score.mopt_env}
\title{compute the simulated score from the
the chain. This is the matrix of the derivative of the
moments with respect to the parameters around the best parameter
value.}
\usage{
  score.mopt_env(me)
}
\description{
  compute the simulated score from the the chain. This is
  the matrix of the derivative of the moments with respect
  to the parameters around the best parameter value.
}

