% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inc.mopt.r
\name{runMOpt}
\alias{runMOpt}
\title{this is the main function, it will run the
optimizer in parallel if called with MPI=TRUE
it will search for the minimum}
\usage{
runMOpt(cf, autoload = TRUE)
}
\description{
this is the main function, it will run the
optimizer in parallel if called with MPI=TRUE
it will search for the minimum
}

